perf(polyfill): polyfill all request animation
