+++ src/components/time-picker/nz-timepicker-inner.component.ts
@@ -11,6 +11,7 @@ import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import * as moment from 'moment';
 import { DropDownAnimation } from '../core/animation/dropdown-animations';
 import { NzLocaleService } from '../locale/index';
+import { reqAnimFrame } from '../core/polyfill/request-animation';
 
 export interface TimeUnitInterface {
   index: number;
@@ -196,9 +197,6 @@ export class NzTimePickerInnerComponent implements OnInit, ControlValueAccessor 
 
   // got from rc-timepicker
   scrollTo(element, to, duration) {
-    const requestAnimationFrame = window.requestAnimationFrame || function requestAnimationFrameTimeout() {
-      return setTimeout(arguments[ 0 ], 10);
-    };
     if (duration <= 0) {
       element.scrollTop = to;
       return;
@@ -206,7 +204,7 @@ export class NzTimePickerInnerComponent implements OnInit, ControlValueAccessor 
     const difference = to - element.scrollTop;
     const perTick = difference / duration * 10;
 
-    requestAnimationFrame(() => {
+    reqAnimFrame(() => {
       element.scrollTop = element.scrollTop + perTick;
       if (element.scrollTop === to) {
         return;
